openid-server/osmclient/requirements.txt
#######################################################################################
# Copyright ETSI Contributors and Others.
#
# Licensed under the Apache License, Version 2.0 (the "License");
# you may not use this file except in compliance with the License.
# You may obtain a copy of the License at
#
#    http://www.apache.org/licenses/LICENSE-2.0
#
# Unless required by applicable law or agreed to in writing, software
# distributed under the License is distributed on an "AS IS" BASIS,
# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or
# implied.
# See the License for the specific language governing permissions and
# limitations under the License.
#######################################################################################
certifi==2022.12.7
    # via requests
charset-normalizer==3.0.1
    # via requests
click==8.1.3
    # via -r requirements.in
idna==3.4
    # via requests
jinja2==3.1.2
    # via -r requirements.in
markupsafe==2.1.2
    # via jinja2
packaging==23.0
    # via -r requirements.in
prettytable==3.6.0
    # via -r requirements.in
pycurl==7.45.2
    # via -r requirements.in
python-magic==0.4.27
    # via -r requirements.in
pyyaml==5.4.1
    # via -r requirements.in
requests==2.28.2
    # via -r requirements.in
urllib3==1.26.14
    # via requests
verboselogs==1.7
    # via -r requirements.in
wcwidth==0.2.6
    # via prettytable
